requirements.txt
# CLI 框架
typer>=0.9.0
rich>=13.0.0
click>=8.1.0

# 数据库 ORM
sqlalchemy>=2.0.0

# 数据分析
pandas>=2.0.0
numpy>=1.24.0

# 可视化
matplotlib>=3.7.0
plotly>=5.15.0

# 配置
pyyaml>=6.0

# 工具
python-dateutil>=2.8.0
tabulate>=0.9.0

# AI / ML（预测模块）
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# AI / LLM（分析模块）
httpx>=0.27.0              # LLM API 调用

# AI / DL（可选，深度学习预测）
# torch>=2.1.0          # 按需安装: pip install torch
